api/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
pydantic==2.10.2
google-generativeai==0.8.3  
httpx==0.28.1
python-dotenv==1.0.1